+++ src/Bootstrap.php
@@ -35,15 +35,23 @@ class Bootstrap {
 	protected static $options;
 
 	/**
+	 * Holds `wp-config.php` file path.
+	 *
+	 * @var $config_path
+	 */
+	protected static $config_path;
+
+	/**
 	 * Constructor.
 	 *
 	 * @param  string $file Main plugin file.
 	 * @return void
 	 */
 	public function __construct( $file ) {
-		$this->file    = $file;
-		$this->dir     = dirname( $file );
-		self::$options = get_site_option( 'wp_debugging', [ 'wp_debug' => '1' ] );
+		$this->file        = $file;
+		$this->dir         = dirname( $file );
+		self::$options     = get_site_option( 'wp_debugging', [ 'wp_debug' => '1' ] );
+		self::$config_path = $this->get_config_path();
 		@ini_set( 'display_errors', 1 );
 	}
 
@@ -55,8 +63,36 @@ class Bootstrap {
 	public function run() {
 		require_once $this->dir . '/vendor/autoload.php';
 		$this->load_hooks();
-		( new Settings( self::$options ) )->load_hooks();
+		( new Settings( self::$options, self::$config_path ) )->load_hooks();
 		\WP_Dependency_Installer::instance()->run( $this->dir );
+	}
+
+	/**
+	 * Get the `wp-config.php` file path.
+	 *
+	 * The config file may reside one level above ABSPATH but is not part of another installation.
+	 *
+	 * @see wp-load.php#L26-L42
+	 *
+	 * @return string $config_path
+	 */
+	public function get_config_path() {
+		$config_path = ABSPATH . 'wp-config.php';
+
+		if ( ! file_exists( $config_path ) ) {
+			if ( @file_exists( dirname( ABSPATH ) . '/wp-config.php' ) && ! @file_exists( dirname( ABSPATH ) . '/wp-settings.php' ) ) {
+				$config_path = dirname( ABSPATH ) . '/wp-config.php';
+			}
+		}
+
+		/**
+		 * Filter the config file path.
+		 *
+		 * @since 2.3.0
+		 *
+		 * @param string $config_path
+		 */
+		return apply_filters( 'wp_debugging_config_path', $config_path );
 	}
 
 	/**
@@ -93,7 +129,7 @@ class Bootstrap {
 	 * @return void
 	 */
 	public function activate() {
-		$config_transformer = new \WPConfigTransformer( ABSPATH . 'wp-config.php' );
+		$config_transformer = new \WPConfigTransformer( self::$config_path );
 		$constants          = [ 'wp_debug_log', 'script_debug', 'savequeries' ];
 		$constants          = array_merge( array_keys( self::$options ), $constants );
 		$config_args        = [
@@ -113,7 +149,7 @@ class Bootstrap {
 	 * @return void
 	 */
 	public function deactivate() {
-		$config_transformer = new \WPConfigTransformer( ABSPATH . 'wp-config.php' );
+		$config_transformer = new \WPConfigTransformer( self::$config_path );
 		$constants          = [ 'wp_debug_log', 'script_debug', 'savequeries', 'wp_debug', 'wp_debug_display', 'wp_disable_fatal_error_handler' ];
 		foreach ( $constants as $constant ) {
 			$config_transformer->remove( 'constant', strtoupper( $constant ) );