+++ CHANGES.md
@@ -1,4 +1,6 @@
 #### [unreleased]
+* look for `wp-config.php` in directory above `ABSPATH`
+* add filter `wp_debugging_config_path` to set non-standard path to `wp-config.php`
 
 #### 2.2.0 / 2019-02-02 🏈
 * initial release on dot org